mw/com: Fix clang-tidy hicpp-signed-bitwise
issue: SWP-239907
GIT_ORIGIN_SPP_REV_ID: cc4648f06c4912726fe1e2ef56f8e42d0fa25db7

+++ score/mw/com/impl/bindings/lola/methods/proxy_instance_identifier.h
@@ -63,7 +63,7 @@ class hash<score::mw::com::impl::lola::ProxyInstanceIdentifier>
         constexpr auto proxy_instance_counter_bit_width =
             std::numeric_limits<decltype(proxy_instance_identifier.proxy_instance_counter)>::digits;
         return std::hash<std::uint64_t>{}((static_cast<std::uint64_t>(proxy_instance_identifier.process_identifier)
-                                           << proxy_instance_counter_bit_width) |
+                                           << static_cast<std::uint64_t>(proxy_instance_counter_bit_width)) |
                                           static_cast<std::uint64_t>(proxy_instance_identifier.proxy_instance_counter));
     }
 };
